Multiply dense values in double, the precision their norms already use
The per-row dense path expands the squared distance as
querySquaredNorm + rowSquaredNorms[i] - 2*dotProduct. Both norms come
from computeSquaredNorm, which widens each value before squaring it and
sums through StableSumAccumulator. The dot product they are subtracted
against multiplied two floats and widened only the result, so one term
of the expansion carried float32 rounding while the other two did not.
The expansion cancels completely for a close pair, which leaves that
rounding as the whole of the answer.

At 512 unit-normalized dimensions a row queried by its own record scored
0.99995166 rather than 1.0, and a pair separated by 1e-3 in one
coordinate was off by the same order. Widening one operand, as
L2Comparator already does for its own squares, brings both within 1e-7.

The matrix multiply keeps float32 accumulation. Its precision is a
property of the kernel, which must agree with OpenBLAS sgemv, and the
exactness the matrix index claims is that it scores every row rather
than pruning the candidate set.

The new test needs a full-length embedding to see any of this: the
existing dense records are three dimensions wide, too few for the
rounding to accumulate. It scores the pair through the distance
definition directly rather than the expansion, so the reference does not
restate the arithmetic under test.

+++ src/main/java/com/uber/ussi/searchablestructure/index/matrix/MatrixIndex.java
@@ -228,7 +228,7 @@ public final class MatrixIndex extends Index {
     int offset = matrixRowIndex * dimension;
     double dotProduct = 0.0d;
     for (int i = 0; i < dimension; ++i) {
-      dotProduct += queryValues[i] * rowMajorValues[offset + i];
+      dotProduct += (double) queryValues[i] * rowMajorValues[offset + i];
     }
     return computeL2SimilarityFromDotProduct(querySquaredNorm, matrixRowIndex, dotProduct);
   }
